Agent: improved template checking after failure
Marvel depends on the marvel index template. This commit makes the agent more diligent in checking the existence of the template after error conditions of settings updates. This is relevent, for example, where a dedicated cluster is being stopped, data folder being cleaned and the cluster brought up. Another issue was found when calls to upload the template connected to the monitoring cluster but returned with a non 2XX status code but no Java Exception was thrown.

Closes #335