[settings/objectEditor] read mapping rather than using service#get
We were previously using service#get() to load a saved object before loading it into the saved object editor. This allowed
us to edit fields that were not actually saved to the object but needed an editor. Now, we are reading the mapping config
off of the SavedObject class so that we don't have to actually load the savedObject.

This fixes https://github.com/elastic/kibana/issues/3914 which was caused by the savedObject class rejecting the #get()
promise due to a missing indexPattern.