Kibana install metadata available as module
The metadata for the current kibana project install that is seeded to
the client-side app via window.__KBN__ is now accessible anywhere via
the immutable ui/metadata singleton. window.__KBN__ now becomes a tested
implementation detail of only a single module in the whole app, whereas
any consumer of the metadata now has an explicit module dependency and
has no way to modify the metadata contents.