Create indexList via field stats for wildcard patterns
Elasticsearch 2.0 introduces the concept of index constraints in the
field stats api, which allows us to dynamically determine which specific
indices need to be queried that match a given wildcard index pattern and
fall within a specific time range.

When we encounter a wildcard index pattern, we now send out a pre-flight
request to the field stats api using the currently selected time range
as index constraints to dynamically build the corresponding index list.