Prevent timepicker day from changing when modifying hours/minutes. Closes #5370.
When a moment date with a timezone is passed to a bootstrap
timepicker the timezone is converted to browser time.  This can cause
the timepicker to show a day that does not match with an input field
used to modify the date string.  On input this will offset the timezone
to prevent the day from changing, and remove the offfset before
outputting.

Fixes #5478