[autoloading] implement autoloading via modules
Currently there is a long list of modules that is autoloaded for every app. This is because a number of angular directives/filters/services are not properly required by the code that needs them and rather than attempt to track down the relationships between every part of the app and the rest autoloading just makes sure that all of the defined modules are included. Since this was implicit it caused new apps which didn't need anything but chrome styles to be really heavy.

This change makes autoloading something you opt into by requiring a ui/autoload/* module. The options include:
 - `ui/autoload/styles` - include the base styles used by Kibana. This includes all of the styles listed in the ui/styles directory.
 - `ui/autoload/directives`
 - `ui/autoload/filters`
 - `ui/autoload/modules` - include angular and several ui modules that one might expect to be loaded by default. This list is hard coded into this file.
 - `ui/autoload/all` - includes all of the above lists

In order to support this change all plugins will likely need to update and include a `ui/autoload/*` module in their public/index.js file.