[devServer] include self-signed certificates for testing
When running the Kibana server in --dev mode the server will use the new `test/dev_certs/server.key/crt` files to provide SSL by default. These files are not secure, and should not be used in production, but are fine for local development. Directions for trusting these certs are linked from the CONTRIBUTING.md file and the new ssl config can be disabled completely by passing `--no-ssl` to the server on startup.