[renderDirective] add support for defining scope bindings
People can use the `<render-directive>` directive to programatically extend views, but passing values to those views is not as easy. To make it easier we will support defining scope bindings just like directives support, and those bindings will be applied to the isolate scope created for the rendered directive.

The `<render-directive>` the majority of angular 1.4s syntax. The differences include:
 1. `=?` is not implemented
 1. Isolate scopes are always created, even is `scope:` is not defined