[uiExports] add uiExport type "injectVars"
In order for plugins to inject default injected vars, they should define an "injectVars" uiExport that will be called once in order to build the default injected vars. This function should return an object with the injectedVars it wishes to add. Plugins are run in a non-deterministic order, so they should not be written to override each other. injectedVars should all have unique names. The injectedVars for the app will override any values provided by "default" injectedVars functions.