Don't use extended_bounds for time filter in date_histogram vis
The date histogram vis currently uses extended_bounds on the date_histogram if a time filter is applied which means that these requests can't be cached since the cache key for the request cache for Elasticsearch is based on the request body. The `extended_bounds` here has no effect on the response of the date histogram aggregation because it also sets minDocCount to 0 so will not return any empty buckets.

This change removes the extended bounds parameter for the time filter so these requests can be cached by Elasticsearch.