[internal] Replace var with let in ui/public U-W (no utils/vislib)
This change was applied to any .js files under directories beginning
with `u` through `w` (other than utils and vislib) in src/ui/public.
There are no `y` or `z` directories. The utils directory isn't included
because the generic change breaks something in it. The vislib directory
isn't included because it had too many changes on its own.

This was an automatic replacement from var to let for any variable
declaration that doubles as the initial assignment. Ultimately we want
most of these to be converted to const, but this commit is so large that
it warrants breaking each step of automation up into its own commit.

For example:

`var foo = 'bar';` becomes `let foo = 'var';`

This was accomplished by replacing:
find: `var ([a-zA-Z_$][0-9a-zA-Z_$]*)(\s+)=`
replace: `let $1$2=`