[functionalTests/common] add #try() method
There are a ton of places in the functional tests where the "default" is set explicitly. This sort of defeats the purpose of a default though, so in my spare time I'll be moving tests away from this pattern.

This change removes the timeout argument from #tryForTime() via a new method, simply called #try(). It also replaces all instances of `common.tryForTime(defaultTimeout, ...)` with calls to this method, which shouldn't have cause any functional changes.