[grunt] Task to rebuild archives with new version
This is useful for when a pre-release build is completely tested and
could be released as-is if it weren't for the pre-release suffix and
commit hash. It will extract the archives in the target, replace the
version, sha, and build numbers with the current working copy's, and
then recreate the archives and shas.