Improve parsing performance
Previously I was using PapaParse's preview option, but it turns
out that does not prevent the library from loading the entire file
if you don't use one of the streaming callbacks. So now we have to
do some extra gymnastics to gather the data per row in the step
callback, manually abort parsing if we haven't hit the end of the file
after a given number of sample lines and then digest the data in the
complete callback.

I also added an extra condition to a watcher that caused parsing to
happen twice.