Show highlighted text color for single "tab" items in navigation
This adds a simple CSS rule to highlight the navigation item to be highlighted as though it's active
without adding the underline.

The rule is very specific, which plays into the discussion in #7364. As a follow-up to this, we should
create a bread crumb element to add that auto-builds these kind of things so that plugin authors do not
need to even know the rules in place. This should help to keep all plugins consistent as well as simplify
the effort to make it work.