Enable additional scripted field languages
This commit allows Kibana scripted fields to be written in any language
that has been enabled for inline scripting in ES. By default, that's
expression and painless in 5.0. Groovy and other languages can be
enabled in the Elasticsearch config. However, I've spent the most time
trying to optimize the experience for painless and expression since
those are the two sandboxed languages we should be encouraging people
to use.

Field type options are limited per language based on what makes sense
for that language and for scripte fields in general. For instance,
expression scripts can only be number fields because expressions only
return numbers. No scripted fields can be geo_points, because the geo
aggregations Kibana uses don't support scripts.

This UI enhancement uses the new script languages API added in the
previous commit to get the list of inline scripting languages
dynamically.